Add files via upload
Each of the rectangular segments make up the 7 segment display. When illuminated, the individual segment forms part of a numerical digit. In other words, the illuminated segments are charged by binary input. 1= on; 0 = off; the binary code is displayed using the segments. The data maps to the display. Each one of the 7 segments in the display has a position based on the svg file we uploaded. We don't know what each of the segments correspond to which position so we must view animation to jot down the positions and numbers.  We can label each segment (a, b, c, d, e, f, g ) which represents each LED segment, which then connects to form the numbers 0-9. User basically sees an array of colored pixels as an image 0-9. What we see are the possibilities to make 0-9.